requirements.txt
Flask==2.0.2
#pandas
#numpy
Pillow
#keras
#tensorflow
Werkzeug
#keras_applications
azure
mrest